Improve fault tolerance
The program would just shutdown after a problem. For certain problems,
for example when you have a bad internet connection and immoscout does
not reply properly (resulting in a JSONDecodeError) we can now handle
this with a message printed to the console, and just retrying the next
minute.

Similarly the chromedriver handling is fault sensitive so we've simply
made this a little more easy. This might cause the bot to not reply to
certain offers, but at least it stays running afterwards to try to apply
for others.

We might consider to store offer-ids of failed attempts in a specific
file so you have a better understanding of which offers failed.